+++ components/led/led_indicator/include/led_indicator.h
@@ -1,44 +1,44 @@
-// Copyright 2022-2024 Espressif Systems (Shanghai) CO LTD
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+/*
+ * SPDX-FileCopyrightText: 2023 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
 
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+#pragma once
 
-#ifndef __LED_INDICATOR_H__
-#define __LED_INDICATOR_H__
+#ifdef __cplusplus
+extern "C" {
+#endif
 
 #include <stdint.h>
 #include <stdbool.h>
-
-#include "freertos/FreeRTOS.h"
-#include "freertos/timers.h"
 #include "esp_err.h"
+#include "led_gpio.h"
+#include "led_ledc.h"
+#include "led_custom.h"
 
 /**
- * @brief led on-off state
+ * @brief LED state: 0-100, only hardware that supports to set brightness can adjust brightness.
  *
  */
-typedef enum {
-    LED_STATE_OFF = 0, /*!< turn off the led */
-    LED_STATE_ON = 1,  /*!< turn on the led */
-} blink_step_state_t;
+enum {
+    LED_STATE_OFF = 0,          /*!< turn off the LED */
+    LED_STATE_25_PERCENT = 64,  /*!< 25% brightness, must support to set brightness */
+    LED_STATE_50_PERCENT = 128,  /*!< 50% brightness, must support to set brightness */
+    LED_STATE_75_PERCENT = 191,  /*!< 75% brightness, must support to set brightness */
+    LED_STATE_ON = UINT8_MAX,   /*!< turn on the LED */
+};
 
 /**
  * @brief actions in this type
  *
  */
 typedef enum {
-    LED_BLINK_STOP = -1, /*!< stop the blink */
-    LED_BLINK_HOLD,      /*!< hold the on-off state */
-    LED_BLINK_LOOP,      /*!< loop from first step */
+    LED_BLINK_STOP = -1,   /*!< stop the blink */
+    LED_BLINK_HOLD,        /*!< hold the on-off state */
+    LED_BLINK_BREATHE,     /*!< breathe state */
+    LED_BLINK_BRIGHTNESS,  /*!< set the brightness */
+    LED_BLINK_LOOP,        /*!< loop from first step */
 } blink_step_type_t;
 
 /**
@@ -47,112 +47,114 @@ typedef enum {
  */
 typedef struct {
     blink_step_type_t type;          /*!< action type in this step */
-    blink_step_state_t on_off;       /*!< hold on or off, set NULL if not LED_BLINK_HOLD*/
+    uint8_t value;                   /*!< hold on or off, set NULL if LED_BLINK_STOP or LED_BLINK_LOOP */
     uint32_t hold_time_ms;           /*!< hold time(ms), set NULL if not LED_BLINK_HOLD,*/
 } blink_step_t;
 
 /**
- * @brief led indicator blink mode, as a member of led_indicator_config_t
+ * @brief LED indicator blink mode, as a member of led_indicator_config_t
  *
  */
 typedef enum {
-    LED_GPIO_MODE,        /*!< blink with max brightness*/
+    LED_GPIO_MODE,         /*!< blink with max brightness*/
+    LED_LEDC_MODE,         /*!< blink with LEDC driver*/
+    LED_CUSTOM_MODE,       /*!< blink with custom driver*/
 } led_indicator_mode_t;
 
 /**
- * @brief led indicator specified configurations, as a arg when create a new indicator
+ * @brief LED indicator specified configurations, as a arg when create a new indicator
  *
  */
 typedef struct {
-    bool off_level;               /*!< gpio level of turn off. 0 if attach led positive side to esp32 gpio pin, 1 if attach led negative side*/
-    led_indicator_mode_t mode;    /*!< led work mode, eg. gpio or pwm mode */
-    blink_step_t **blink_lists;   /*!< user defined led blink lists */
-    uint16_t blink_list_num;      /*!< number of blink lists */
+    led_indicator_mode_t mode;                  /*!< LED work mode, eg. GPIO or pwm mode */
+    union {
+        led_indicator_gpio_config_t *led_indicator_gpio_config;     /*!< LED GPIO configuration */
+        led_indicator_ledc_config_t *led_indicator_ledc_config;     /*!< LED LEDC configuration */
+        led_indicator_custom_config_t *led_indicator_custom_config; /*!< LED custom configuration */
+    }; /**< LED configuration */
+    blink_step_t const **blink_lists;           /*!< user defined LED blink lists */
+    uint16_t blink_list_num;                    /*!< number of blink lists */
 } led_indicator_config_t;
 
-/**
- * @brief The blink type with smaller index has the higher priority
- * eg. BLINK_FACTORY_RESET priority is higher than BLINK_UPDATING
- */
-typedef enum {
-    BLINK_FACTORY_RESET,           /*!< restoring factory settings */
-    BLINK_UPDATING,                /*!< updating software */
-    BLINK_CONNECTED,               /*!< connected to AP (or Cloud) succeed */
-    BLINK_PROVISIONED,             /*!< provision done */
-    BLINK_CONNECTING,              /*!< connecting to AP (or Cloud) */
-    BLINK_RECONNECTING,            /*!< reconnecting to AP (or Cloud), if lose connection */
-    BLINK_PROVISIONING,            /*!< provisioning */
-    BLINK_MAX,                     /*!< INVALID type */
-} led_indicator_blink_type_t;
+typedef void *led_indicator_handle_t; /*!< LED indicator operation handle */
 
 /**
- * @brief Get the number of sample led blinking modes.
+ * @brief create a LED indicator instance with GPIO number and configuration
  *
- * @return uint16_t
+ * @param config configuration of the LED, eg. GPIO level when LED off
+ * @return led_indicator_handle_t handle of the LED indicator, NULL if create failed.
  */
-uint16_t led_indicator_get_sample_lists_num(void);
+led_indicator_handle_t led_indicator_create(const led_indicator_config_t *config);
 
 /**
- * @brief Get example led blinking list pointer.
+ * @brief get the handle of created led_indicator with hardware data
  *
- * @return blink_step_t const**
+ * @param hardware_data user hardware data for LED
+ * @return led_indicator_handle_t handle of the created LED indicator, NULL if not created.
  */
-blink_step_t const **led_indicator_get_sample_lists(void);
-
-
-typedef void *led_indicator_handle_t; /*!< led indicator operation handle */
+led_indicator_handle_t led_indicator_get_handle(void *hardware_data);
 
 /**
- * @brief create a led indicator instance with gpio number and configuration
+ * @brief delete the LED indicator and release resource
  *
- * @param io_num gpio number of the led
- * @param config configuration of the led, eg. gpio level when led off
- * @return led_indicator_handle_t handle of the led indicator, NULL if create failed.
- */
-led_indicator_handle_t led_indicator_create(int io_num, const led_indicator_config_t *config);
-
-/**
- * @brief get the handle of created led_indicator with gpio number
- *
- * @param io_num  gpio number of the led
- * @return led_indicator_handle_t handle of the created led indicator, NULL if not created
- */
-led_indicator_handle_t led_indicator_get_handle(int io_num);
-
-/**
- * @brief delete the led indicator and release resource
- *
- * @param p_handle pointer to led indicator handle
+ * @param handle pointer to LED indicator handle
  * @return esp_err_t
- *     - ESP_ERR_INVALID_ARG   if parameter is invalid
- *     - ESP_FAIL Fail
+ *     - ESP_ERR_INVALID_ARG if parameter is invalid
  *     - ESP_OK Success
+ *     - ESP_FAIL Delete fail
  */
-esp_err_t led_indicator_delete(led_indicator_handle_t *p_handle);
+esp_err_t led_indicator_delete(led_indicator_handle_t handle);
 
 /**
- * @brief start a new blink_type on the led indicator. if mutiple blink_type started simultaneously,
+ * @brief start a new blink_type on the LED indicator. if mutiple blink_type started simultaneously,
  * it will be executed according to priority.
  *
- * @param handle led indicator handle
+ * @param handle LED indicator handle
  * @param blink_type predefined blink type
  * @return esp_err_t
- *     - ESP_ERR_INVALID_ARG   if parameter is invalid
+ *     - ESP_ERR_INVALID_ARG if parameter is invalid
  *     - ESP_ERR_NOT_FOUND no predefined blink_type found
  *     - ESP_OK Success
  */
-esp_err_t led_indicator_start(led_indicator_handle_t handle, led_indicator_blink_type_t blink_type);
+esp_err_t led_indicator_start(led_indicator_handle_t handle, int blink_type);
 
 /**
  * @brief stop a blink_type. you can stop a blink_type at any time, no matter it is executing or waiting to be executed.
  *
- * @param handle led indicator handle
+ * @param handle LED indicator handle
  * @param blink_type predefined blink type
  * @return esp_err_t
- *     - ESP_ERR_INVALID_ARG   if parameter is invalid
+ *     - ESP_ERR_INVALID_ARG if parameter is invalid
  *     - ESP_ERR_NOT_FOUND no predefined blink_type found
  *     - ESP_OK Success
  */
-esp_err_t led_indicator_stop(led_indicator_handle_t handle, led_indicator_blink_type_t blink_type);
+esp_err_t led_indicator_stop(led_indicator_handle_t handle, int blink_type);
 
+/**
+ * @brief Immediately execute an action of any priority. Until the action is executed, or call led_indicator_preempt_stop().
+ *
+ * @param handle LED indicator handle
+ * @param blink_type predefined blink type
+ * @return esp_err_t
+ *     - ESP_OK Success
+ *     - ESP_FAIL Fail
+ *     - ESP_ERR_INVALID_ARG if parameter is invalid
+ */
+esp_err_t led_indicator_preempt_start(led_indicator_handle_t handle, int blink_type);
+
+/**
+ * @brief Stop the current preemptive action.
+ *
+ * @param handle LED indicator handle
+ * @param blink_type predefined blink type
+ * @return esp_err_t
+ *     - ESP_OK Success
+ *     - ESP_FAIL Fail
+ *     - ESP_ERR_INVALID_ARG if parameter is invalid
+ */
+esp_err_t led_indicator_preempt_stop(led_indicator_handle_t handle, int blink_type);
+
+#ifdef __cplusplus
+}
 #endif
+