+++ components/led/led_indicator/CHANGELOG.md
@@ -1,5 +1,16 @@
 # ChangeLog
 
+## v0.3.0 - 2023-1-31
+
+### Enhancements:
+
+* LED INDICATOR:
+  * Support customer mode: can customize multiple drivers
+  * Support create LED by LEDC driver
+  * Support breathe blink
+  * Support set brightness
+  * Support jump in line
+
 ## v0.2.0 - 2022-12-28
 
 ### Enhancements: